Conflicts:
	Core/Config.cpp
	UI/ProfilerDraw.cpp
	ext/native/profiler/profiler.h

+++ Core/MIPS/ARM64/Arm64CompBranch.cpp
@@ -128,26 +128,56 @@ void Arm64Jit::BranchRSRTComp(MIPSOpcode op, CCFlags cc, bool likely)
 
 		// We might be able to flip the condition (EQ/NEQ are easy.)
 		const bool canFlip = cc == CC_EQ || cc == CC_NEQ;
-
-		// TODO ARM64: Optimize for immediates other than zero
-		if (rt == 0) {
-			gpr.MapIn(rs);
-			CMP(gpr.R(rs), 0);
-		} else {
-			gpr.MapInIn(rs, rt);
-			CMP(gpr.R(rs), gpr.R(rt));
-		}
+		const bool rsIsZero = gpr.IsImm(rs) && gpr.GetImm(rs) == 0;
+		const bool rtIsZero = gpr.IsImm(rt) && gpr.GetImm(rt) == 0;
 
 		Arm64Gen::FixupBranch ptr;
-		if (!likely) {
-			if (!delaySlotIsNice)
-				CompileDelaySlot(DELAYSLOT_SAFE_FLUSH);
-			else
-				FlushAll();
-			ptr = B(cc);
-		} else {
+		if ((likely || delaySlotIsNice) && (rsIsZero || rtIsZero) && canFlip) {
+			// Special case, we can just use CBZ/CBNZ directly.
+			MIPSGPReg r = rsIsZero ? rt : rs;
+			gpr.MapReg(r);
+			// Flush should keep r in the same armreg.
+			ARM64Reg ar = gpr.R(r);
 			FlushAll();
-			ptr = B(cc);
+			if (cc == CC_EQ) {
+				ptr = CBZ(ar);
+			} else {
+				ptr = CBNZ(ar);
+			}
+		} else {
+			u32 val;
+			bool shift;
+			if (gpr.IsImm(rt) && IsImmArithmetic(gpr.GetImm(rt), &val, &shift)) {
+				gpr.MapReg(rs);
+				CMP(gpr.R(rs), val, shift);
+			} else if (gpr.IsImm(rt) && IsImmArithmetic((u64)(s64)-(s32)gpr.GetImm(rt), &val, &shift)) {
+				gpr.MapReg(rs);
+				CMN(gpr.R(rs), val, shift);
+			} else if (gpr.IsImm(rs) && IsImmArithmetic(gpr.GetImm(rs), &val, &shift) && canFlip) {
+				gpr.MapReg(rt);
+				CMP(gpr.R(rt), val, shift);
+			} else if (gpr.IsImm(rs) && IsImmArithmetic((u64)(s64)-(s32)gpr.GetImm(rs), &val, &shift) && canFlip) {
+				gpr.MapReg(rt);
+				CMN(gpr.R(rt), val, shift);
+			} else {
+				gpr.MapInIn(rs, rt);
+				CMP(gpr.R(rs), gpr.R(rt));
+			}
+
+			if (!likely) {
+				if (!delaySlotIsNice)
+					CompileDelaySlot(DELAYSLOT_SAFE_FLUSH);
+				else
+					FlushAll();
+				ptr = B(cc);
+			} else {
+				FlushAll();
+				ptr = B(cc);
+			}
+		}
+
+		if (likely) {
+			// Only executed when taking the branch.
 			CompileDelaySlot(DELAYSLOT_FLUSH);
 		}
 
@@ -234,7 +264,6 @@ void Arm64Jit::BranchRSZeroComp(MIPSOpcode op, CCFlags cc, bool andLink, bool li
 		if (!likely && delaySlotIsNice)
 			CompileDelaySlot(DELAYSLOT_NICE);
 
-		// TODO: Maybe we could use BZ here?
 		gpr.MapReg(rs);
 		CMP(gpr.R(rs), 0);
 
@@ -324,19 +353,24 @@ void Arm64Jit::BranchFPFlag(MIPSOpcode op, CCFlags cc, bool likely) {
 	if (!likely && delaySlotIsNice)
 		CompileDelaySlot(DELAYSLOT_NICE);
 
-	// TODO: Maybe we could use TBZ here?
 	gpr.MapReg(MIPS_REG_FPCOND);
-	TSTI2R(gpr.R(MIPS_REG_FPCOND), 1, SCRATCH1);
 	Arm64Gen::FixupBranch ptr;
-	if (!likely) {
-		if (!delaySlotIsNice)
-			CompileDelaySlot(DELAYSLOT_SAFE_FLUSH);
-		else
-			FlushAll();
-		ptr = B(cc);
-	} else {
+	if (likely || delaySlotIsNice) {
+		// FlushAll() won't actually change the reg.
+		ARM64Reg ar = gpr.R(MIPS_REG_FPCOND);
 		FlushAll();
+		if (cc == CC_EQ) {
+			ptr = TBZ(ar, 0);
+		} else {
+			ptr = TBNZ(ar, 0);
+		}
+	} else {
+		TSTI2R(gpr.R(MIPS_REG_FPCOND), 1, SCRATCH1);
+		CompileDelaySlot(DELAYSLOT_SAFE_FLUSH);
 		ptr = B(cc);
+	}
+
+	if (likely) {
 		CompileDelaySlot(DELAYSLOT_FLUSH);
 	}
 
@@ -385,28 +419,26 @@ void Arm64Jit::BranchVFPUFlag(MIPSOpcode op, CCFlags cc, bool likely) {
 
 	int imm3 = (op >> 18) & 7;
 
-	// TODO: Maybe could use TBZ?
 	gpr.MapReg(MIPS_REG_VFPUCC);
-	TSTI2R(gpr.R(MIPS_REG_VFPUCC), 1 << imm3, SCRATCH1);
-
 	Arm64Gen::FixupBranch ptr;
-	js.inDelaySlot = true;
-	if (!likely)
-	{
-		if (!delaySlotIsNice && !delaySlotIsBranch)
-			CompileDelaySlot(DELAYSLOT_SAFE_FLUSH);
-		else
-			FlushAll();
-		ptr = B(cc);
-	}
-	else
-	{
+	if (likely || delaySlotIsNice || delaySlotIsBranch) {
+		// FlushAll() won't actually change the reg.
+		ARM64Reg ar = gpr.R(MIPS_REG_VFPUCC);
 		FlushAll();
+		if (cc == CC_EQ) {
+			ptr = TBZ(ar, imm3);
+		} else {
+			ptr = TBNZ(ar, imm3);
+		}
+	} else {
+		TSTI2R(gpr.R(MIPS_REG_VFPUCC), 1 << imm3, SCRATCH1);
+		CompileDelaySlot(DELAYSLOT_SAFE_FLUSH);
 		ptr = B(cc);
-		if (!delaySlotIsBranch)
-			CompileDelaySlot(DELAYSLOT_FLUSH);
 	}
-	js.inDelaySlot = false;
+
+	if (likely && !delaySlotIsBranch) {
+		CompileDelaySlot(DELAYSLOT_FLUSH);
+	}
 
 	// Take the branch
 	WriteExit(targetAddr, js.nextExit++);
@@ -504,7 +536,7 @@ void Arm64Jit::Comp_JumpReg(MIPSOpcode op)
 		delaySlotIsNice = false;
 	CONDITIONAL_NICE_DELAYSLOT;
 
-	ARM64Reg destReg = OTHERTEMPREG;
+	ARM64Reg destReg = INVALID_REG;
 	if (IsSyscall(delaySlotOp)) {
 		gpr.MapReg(rs);
 		MovToPC(gpr.R(rs));  // For syscall to be able to return.
@@ -542,7 +574,9 @@ void Arm64Jit::Comp_JumpReg(MIPSOpcode op)
 		destReg = gpr.R(rs);  // Safe because FlushAll doesn't change any regs
 		FlushAll();
 	} else {
-		// Delay slot - this case is very rare, might be able to free up R8.
+		// Since we can't be in a delay slot, should be safe to steal FLAGTEMPREG for a temp reg.
+		// It will be saved, even if a function is called.
+		destReg = FLAGTEMPREG;
 		gpr.MapReg(rs);
 		MOV(destReg, gpr.R(rs));
 		if (andLink)
